+++ src/strategy/raids/gruulslair/RaidGruulsLairHelpers.cpp
@@ -124,7 +124,7 @@ bool IsPositionSafe(PlayerbotAI* botAI, Player* bot, Position pos)
 bool TryGetNewSafePosition(PlayerbotAI* botAI, Player* bot, Position& outPos)
 {
     const float SEARCH_RADIUS = 30.0f;
-    const int NUM_POSITIONS = 32;
+    const uint8 NUM_POSITIONS = 32;
 
     outPos = { bot->GetPositionX(), bot->GetPositionY(), bot->GetPositionZ() };
     if (IsPositionSafe(botAI, bot, outPos))
